Updated to 1.20.4

+++ src/main/java/net/fabricmc/renew_auto_plus/WastesDimensionEffect.java
@@ -10,7 +10,7 @@ public class WastesDimensionEffect extends DimensionEffects {
 
     @Override
     public Vec3d adjustFogColor(Vec3d color, float sunHeight) {
-        return color.multiply(sunHeight * 0.94f + 0.06f, sunHeight * 0.94f + 0.06f, sunHeight * 0.91f + 0.09f);
+        return color.multiply((double)(sunHeight * 0.94f + 0.06f), (double)(sunHeight * 0.94f + 0.06f), (double)(sunHeight * 0.91f + 0.09f));
     }
 
     @Override
